# Improve marshaling/unmarshaling performance for NodeStream calltypes