--- a/cmd/protoc-gen-gorums/dev/encoding.go
+++ b/cmd/protoc-gen-gorums/dev/encoding.go
@@ -4,10 +4,17 @@ import (
 	"fmt"
 
 	"github.com/relab/gorums/ordering"
+	"google.golang.org/grpc/encoding"
 	"google.golang.org/protobuf/encoding/protowire"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/reflect/protoreflect"
 )
+
+const gorumsContentType = "gorums"
+
+func init() {
+	encoding.RegisterCodec(newGorumsCodec())
+}
 
 type gorumsMessage struct {
 	metadata *ordering.Metadata
@@ -31,6 +38,10 @@ func newGorumsCodec() *gorumsCodec {
 	}
 }
 
+func (c gorumsCodec) Name() string {
+	return gorumsContentType
+}
+
 func (c gorumsCodec) Marshal(m interface{}) (b []byte, err error) {
 	switch msg := m.(type) {
 	case *gorumsMessage:
@@ -42,17 +53,21 @@ func (c gorumsCodec) Marshal(m interface{}) (b []byte, err error) {
 	}
 }
 
+// gorumsMarshal marshals a metadata and a data message into a single byte slice.
 func (c gorumsCodec) gorumsMarshal(msg *gorumsMessage) (b []byte, err error) {
-	md, err := c.marshaler.Marshal(msg.metadata)
+	mdSize := c.marshaler.Size(msg.metadata)
+	b = protowire.AppendVarint(b, uint64(mdSize))
+	b, err = c.marshaler.MarshalAppend(b, msg.metadata)
 	if err != nil {
 		return nil, err
 	}
-	b = protowire.AppendBytes(b, md)
-	data, err := c.marshaler.Marshal(msg.message)
+
+	msgSize := c.marshaler.Size(msg.message)
+	b = protowire.AppendVarint(b, uint64(msgSize))
+	b, err = c.marshaler.MarshalAppend(b, msg.message)
 	if err != nil {
 		return nil, err
 	}
-	b = protowire.AppendBytes(b, data)
 	return b, nil
 }
 
@@ -67,6 +82,7 @@ func (c gorumsCodec) Unmarshal(b []byte, m interface{}) (err error) {
 	}
 }
 
+// gorumsUnmarshal unmarshals a metadata and a data message from a byte slice.
 func (c gorumsCodec) gorumsUnmarshal(b []byte, msg *gorumsMessage) (err error) {
 	mdBuf, mdLen := protowire.ConsumeBytes(b)
 	err = c.unmarshaler.Unmarshal(mdBuf, msg.metadata)