--- a/cmd/protoc-gen-gorums/dev/node.go
+++ b/cmd/protoc-gen-gorums/dev/node.go
@@ -10,7 +10,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/relab/gorums/ordering"
 	"google.golang.org/grpc"
 )
 
@@ -37,7 +36,7 @@ type Node struct {
 func (n *Node) createOrderedStream(rq *receiveQueue, opts managerOptions) {
 	n.orderedNodeStream = &orderedNodeStream{
 		receiveQueue: rq,
-		sendQ:        make(chan *ordering.Message, opts.sendBuffer),
+		sendQ:        make(chan *gorumsMessage, opts.sendBuffer),
 		node:         n,
 		backoff:      opts.backoff,
 		rand:         rand.New(rand.NewSource(time.Now().UnixNano())),