--- a/cmd/protoc-gen-gorums/dev/server.go
+++ b/cmd/protoc-gen-gorums/dev/server.go
@@ -118,7 +118,6 @@ func WithGRPCServerOptions(opts ...grpc.ServerOption) ServerOption {
 type GorumsServer struct {
 	srv        *orderingServer
 	grpcServer *grpc.Server
-	opts       serverOptions
 }
 
 // NewGorumsServer returns a new instance of GorumsServer.