--- a/ordering/ordering.proto
+++ b/ordering/ordering.proto
@@ -5,8 +5,9 @@ option go_package = "github.com/relab/gorums/ordering";
 
 service Gorums { rpc NodeStream(stream Message) returns (stream Message); }
 
-message Message {
-  uint64 ID = 1;
+message Message {}
+
+message Metadata {
+  uint64 MessageID = 1;
   int32 MethodID = 2;
-  bytes Data = 3;
 }