… benchmark/benchmark.pb.go
… benchmark/benchmark_gorums.pb.go
… benchmark/benchmark_grpc.pb.go
… benchmark/server.go
… cmd/protoc-gen-gorums/dev/encoding_test.go
… cmd/protoc-gen-gorums/dev/zorums.pb.go
… cmd/protoc-gen-gorums/dev/zorums_ordered_gorums.pb.go
… cmd/protoc-gen-gorums/gengorums/template_ordered_qc.go
… cmd/protoc-gen-gorums/gengorums/template_ordered_rpc.go
… cmd/protoc-gen-gorums/gengorums/template_unicast.go
… go.sum
… gorums.pb.go
… internal/correctable/opts.pb.go
… internal/ordering/opts.pb.go
… internal/testprotos/calltypes/async/async_gorums.pb.go
… internal/testprotos/calltypes/importmsgs/importmsgs_gorums.pb.go
… internal/testprotos/calltypes/multicast/multicast_gorums.pb.go
… internal/testprotos/calltypes/quorumcall/quorumcall_gorums.pb.go
… internal/testprotos/calltypes/zorums/zorums_gorums.pb.go
… ordering/ordering.pb.go
… ordering/ordering_grpc.pb.go
… tests/Makefile
… tests/message_encoding/encoding_test.go
… tests/message_encoding/message_encoding.pb.go
… tests/message_encoding/message_encoding.proto
… tests/message_encoding/message_encoding_gorums.pb.go
… tests/ordering/order.pb.go
… tests/ordering/order_gorums.pb.go
… tests/qf/qf.pb.go
… tests/qf/qf_gorums.pb.go