This PR implements a gRPC [`Codec`](https://pkg.go.dev/google.golang.org/grpc/encoding?tab=doc#Codec) that performs fast marshaling and unmarshaling of a gorums message consisting of metadata and an application-specific data message. This is accomplished by marshaling both the metadata and data messages onto the same byte slice using [`MarshalAppend`](https://pkg.go.dev/google.golang.org/protobuf/proto?tab=doc#MarshalOptions.MarshalAppend) and prefixing each marshaled message with its length. Unmarshaling is simply done by reading the length of each message and unmarshaling them.

This approach offers better performance than the old approach because it avoids copying the data message more than once. See #78 for more details.

Fixes #78