@Raytar 2020-06-24
Benchmark of this new approach vs the old:
```
goos: linux
goarch: amd64
pkg: github.com/relab/gorums/tests/message_encoding
BenchmarkMetadataWithEmbeddedMessage/Marshal/10-8        1000000              2235 ns/op             240 B/op          5 allocs/op
BenchmarkMetadataWithEmbeddedMessage/Marshal/100-8        415006              2829 ns/op             432 B/op          5 allocs/op
BenchmarkMetadataWithEmbeddedMessage/Marshal/1000-8       337478              4005 ns/op            2224 B/op          5 allocs/op
BenchmarkMetadataWithEmbeddedMessage/Marshal/10000-8       65666             15400 ns/op           20656 B/op          5 allocs/op
BenchmarkMetadataWithEmbeddedMessage/Marshal/100000-8       8941            133444 ns/op          213168 B/op          5 allocs/op
BenchmarkMetadataWithEmbeddedMessage/Unmarshal/10-8       369932              2779 ns/op             240 B/op          6 allocs/op
BenchmarkMetadataWithEmbeddedMessage/Unmarshal/100-8      448782              2891 ns/op             432 B/op          6 allocs/op
BenchmarkMetadataWithEmbeddedMessage/Unmarshal/1000-8     251708              4706 ns/op            2240 B/op          6 allocs/op
BenchmarkMetadataWithEmbeddedMessage/Unmarshal/10000-8     64753             16017 ns/op           20672 B/op          6 allocs/op
BenchmarkMetadataWithEmbeddedMessage/Unmarshal/100000-8     9732            122785 ns/op          213184 B/op          6 allocs/op
BenchmarkGorumsCodec/Marshal/10-8                         609368              1832 ns/op             184 B/op          4 allocs/op
BenchmarkGorumsCodec/Marshal/100-8                        588836              2003 ns/op             280 B/op          4 allocs/op
BenchmarkGorumsCodec/Marshal/1000-8                       445206              2631 ns/op            1176 B/op          4 allocs/op
BenchmarkGorumsCodec/Marshal/10000-8                      125390              8403 ns/op           10392 B/op          4 allocs/op
BenchmarkGorumsCodec/Marshal/100000-8                      17895             65130 ns/op          106648 B/op          4 allocs/op
BenchmarkGorumsCodec/Unmarshal/10-8                       630361              1966 ns/op             192 B/op          5 allocs/op
BenchmarkGorumsCodec/Unmarshal/100-8                      593460              2616 ns/op             288 B/op          5 allocs/op
BenchmarkGorumsCodec/Unmarshal/1000-8                     363512              3448 ns/op            1200 B/op          5 allocs/op
BenchmarkGorumsCodec/Unmarshal/10000-8                    108781              9518 ns/op           10416 B/op          5 allocs/op
BenchmarkGorumsCodec/Unmarshal/100000-8                    19300             59640 ns/op          106672 B/op          5 allocs/op
```