@Raytar 2020-06-24
Quorum call benchmarks of the new vs the old:

```
New 0 payload:

Benchmark      Throughput         Latency    Std.dev    Client+Servers                     
OrderedQC      3901.46 ops/sec    0.26 ms    0.26 ms    1647 B/op         54 allocs/op     
UnorderedQC    2243.68 ops/sec    0.44 ms    0.45 ms    8743 B/op         168 allocs/op    

Old 0 payload: 
                    
OrderedQC      3915.95 ops/sec    0.25 ms    0.26 ms    1607 B/op         49 allocs/op     
UnorderedQC    2245.67 ops/sec    0.44 ms    0.48 ms    8683 B/op         166 allocs/op    

New 1M payload:  
               
OrderedQC      149.72 ops/sec    6.68 ms    0.82 ms    6199415 B/op      225 allocs/op    
UnorderedQC    142.94 ops/sec    6.99 ms    1.00 ms    6215872 B/op      353 allocs/op    

Old 1M payload:     
            
OrderedQC      118.64 ops/sec    8.43 ms    2.68 ms    10477961 B/op     240 allocs/op    
UnorderedQC    139.99 ops/sec    7.14 ms    1.47 ms    6213726 B/op      352 allocs/op    
```